Refactor core into package modules
